+++ Editor/DisplayXRCameraEditor.cs
@@ -17,6 +17,7 @@ namespace DisplayXR.Editor
 
         void OnEnable()
         {
+            if (target == null) return;
             m_IpdFactor = serializedObject.FindProperty("ipdFactor");
             m_ParallaxFactor = serializedObject.FindProperty("parallaxFactor");
             m_InvConvergenceDistance = serializedObject.FindProperty("invConvergenceDistance");